--- src/components/Footer.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Footer = () => (
-	<div className="container">
-		<p> Reserved Rights FranksDomain </p>
-	</div>
-)
-
-export default Footer;
