+++ src/components/Home.js
@@ -35,21 +35,21 @@ class Home extends Component {
 		window.addEventListener('scroll', this.handleScroll)
 	}
 
-	handleScroll = (event) => {
-		const {page} = this.state
-	    const estaAbajo = document.body.scrollHeight === window.scrollY + window.innerHeight;
+	// handleScroll = (event) => {
+	// 	const {page} = this.state
+	//     const estaAbajo = document.body.scrollHeight === window.scrollY + window.innerHeight;
 
-	    if(estaAbajo) {	
-			this.api.getCharacters(page).then(res => {
-				this.setState({
-					loading:false, 
-					characters: [...this.state.characters, ...res],
-					page: this.state.page+1
-				})
-			})
+	//     if(estaAbajo) {	
+	// 		this.api.getCharacters(page).then(res => {
+	// 			this.setState({
+	// 				loading:false, 
+	// 				characters: [...this.state.characters, ...res],
+	// 				page: this.state.page+1
+	// 			})
+	// 		})
 
-	    }
-  	}
+	//     }
+ //  	}
 
 	componentWillUnmount(){
 		window.removeEventListener('scroll', this.handleScroll)
@@ -64,7 +64,7 @@ class Home extends Component {
 				    <div className="py-5 bg-light">
 				        <div className="container">
 				            <section className="items-section">
-				                <h5 className="items-section-title">Comics Populares <NavLink to="/characters">Ver todas</NavLink></h5>
+				                <h5 className="items-section-title">Comics mas Populares <NavLink to="/characters">Ver todas</NavLink></h5>
 
 				                <div className="items-section-body">
 				                    {loading && <img src="https://i.imgur.com/EH9HF6h.gif" width={200}/>}
