+++ src/components/SearchBox.js
@@ -3,9 +3,9 @@ import '../App.css';
 
 const SearchBox = () => (
 	<div className="container">
-		<form className="form-group" action="">
+		<form className="form-group">
 	     	<input className="form-control" type="text" placeholder="Buscar Película o Serie" aria-label="Search" />
-	     	<button type="submit">Buscar</button>
+	     	<button type="submit" onSubmit={(e)=>e.preventDefault()}>Buscar</button>
 	    </form>
     </div>            
 	)
